src/main/java/ru/nethology/domain/Radio.java
package ru.nethology.domain;

public class Radio {
    private int radioStationNumber;
    private int currentVolume;

    public int getRadioStationNumber() {
        return radioStationNumber;
    }

    public void setRadioStationNumber(int radioStationNumber) {
        if (radioStationNumber <= 0) {
            return;
        }
        if (radioStationNumber > 9) {
            return;
        }
        this.radioStationNumber = radioStationNumber;
    }

    public int getCurrentVolume() {
        return currentVolume;
    }

    public void setCurrentVolume(int currentVolume) {
        this.currentVolume = currentVolume;
    }

    public void nextStation() {
        if (radioStationNumber < 9) {
            radioStationNumber = radioStationNumber + 1;
        }
        if (radioStationNumber == 9) {
            radioStationNumber = 0;
        }
    }

    public void previousStation() {
        if (radioStationNumber > 0) {
            radioStationNumber = radioStationNumber - 1;
        }
        if (radioStationNumber == 0) {
            radioStationNumber = 9;
        }
    }

    public void increaseVolume() {
        if (currentVolume < 10) {
            currentVolume = currentVolume + 1;
        }
    }

    public void volumeDown() {
        if (currentVolume > 0) {
            currentVolume = currentVolume - 1;
        }
    }

}
